[core] Fix database creation

+++ core/src/main/java/org/lsposed/lspd/service/ConfigFileManager.java
@@ -102,22 +102,42 @@ class ConfigFileManager {
     static void migrateOldConfig(ConfigManager configManager) {
         var miscPath = new File(basePath, "misc_path");
         var enableResources = new File(configDirPath, "enable_resources");
+        var verboseLog = new File(configDirPath, "verbose_log");
+        boolean migrationFailed = false;
 
         if (miscPath.exists()) {
             try {
                 var s = "/data/misc/" + readText(miscPath);
+                Log.d(TAG, "misc path: " + s);
                 configManager.updateModulePrefs("lspd", 0, "config", "misc_path", s);
-                miscPath.delete();
-            } catch (IOException ignored) {
+            } catch (IOException e) {
+                migrationFailed = true;
+                Log.e(TAG, Log.getStackTraceString(e));
             }
+            if (!migrationFailed) miscPath.delete();
         }
         if (enableResources.exists()) {
             try {
+                migrationFailed = false;
                 var s = readText(enableResources);
                 var i = Integer.parseInt(s);
                 configManager.updateModulePrefs("lspd", 0, "config", "enable_resources", i == 1);
-                enableResources.delete();
-            } catch (IOException ignored) {
+            } catch (IOException e) {
+                migrationFailed = true;
+                Log.e(TAG, Log.getStackTraceString(e));
+            }
+            if (!migrationFailed) enableResources.delete();
+            if (verboseLog.exists()) {
+                try {
+                    migrationFailed = false;
+                    var s = readText(verboseLog);
+                    var i = Integer.parseInt(s);
+                    configManager.updateModulePrefs("lspd", 0, "config", "enable_verbose_log", i == 1);
+                } catch (IOException e) {
+                    migrationFailed = true;
+                    Log.e(TAG, Log.getStackTraceString(e));
+                }
+                if (!migrationFailed) verboseLog.delete();
             }
         }
     }
